add method to verify the user role
