+++ packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="AWSSDK.Core" version="3.7.300.13" targetFramework="net48" />
+  <package id="AWSSDK.Core" version="3.7.302.7" targetFramework="net48" />
+  <package id="AWSSDK.IdentityManagement" version="3.7.300.47" targetFramework="net48" />
   <package id="AWSSDK.S3" version="3.7.304.2" targetFramework="net48" />
   <package id="AWSSDK.SageMaker" version="3.7.305" targetFramework="net48" />
   <package id="AWSSDK.SageMakerRuntime" version="3.7.301.2" targetFramework="net48" />
